smoke2

+++ Trains3D/Engine/Isometric/EGCameraIso.m
@@ -23,7 +23,7 @@ static ODClassType* _EGCameraIso_type;
     if(self) {
         _tilesOnScreen = tilesOnScreen;
         _center = center;
-        _eyeDirection = EGVec3Make(1.0, -1.0, 0.5);
+        _eyeDirection = EGVec3Make(1.0, -1.0, ((float)(0.5)));
     }
     
     return self;
