+++ Trains3D/Trains3D/View/TRLevelView.m
@@ -34,7 +34,7 @@ static ODClassType* _TRLevelView_type;
         _cityView = [TRCityView cityView];
         _railroadView = [TRRailroadView railroadView];
         _trainView = [TRTrainView trainView];
-        _environment = [EGEnvironment environmentWithAmbientColor:EGColorMake(0.4, 0.4, 0.4, 1.0) lights:(@[[EGDirectLight directLightWithColor:EGColorMake(1.0, 1.0, 1.0, 1.0) direction:EGVec3Make(-0.2, 0.2, -0.5)]])];
+        _environment = [EGEnvironment environmentWithAmbientColor:EGColorMake(0.4, 0.4, 0.4, 1.0) lights:(@[[EGDirectLight directLightWithColor:EGColorMake(1.0, 1.0, 1.0, 1.0) direction:EGVec3Make(((float)(-0.2)), ((float)(0.2)), ((float)(-0.5)))]])];
         _camera = [EGCameraIso cameraIsoWithTilesOnScreen:_level.map.size center:EGPointMake(0.0, 0.0)];
     }
